fix(region): climc add base projects options
